Fixes the issues with holding/unholding customer shipments
The documentation for the API endpoint specified that we needed to
provide a hash to hold/unhold a customer shipment. Upon further investigation
I found that we need to provide a string to both API endpoints with
the reason for holding/unholding the customer shipment.

This ensures that we're sending the expected payload to Fulfil and
that we can hold/unhold customer shipments.