Fixes the nested relation assignment bug
When you're fetching nested data from Fulfil, they don't allow you to
query the ID of the nested resource. Why? No idea, but they have provided
a way around this problem: they return the ID under the name of the resource.

For example: a line item of a sales order.

When querying the name and the id of the sales order of a line item, the
return value from Fulfil looks like this:

{ "sale" => 10, "sale.number" => "CODE-123" }

Before this change, we weren't able to handle this. We would simply only
keep the "sale.number" and dismiss the "sale" key itself. This resulted
in the following response value:

{ "sale" => { "name" => "CODE-123" } }

With these changes, we are able to pick up that ID of the sale. Now, the
result from the Fulfil API is:

{ "sale" => { "id" => 10, "name" => "CODE-123" } }