Fix: extracting string IDs in nested resources
In the past, we already solved the issue whereby Fulfil returns
the ID of a nested resource as the root key.

However, we only solved it whenever the ID of a nested resource
was an integer. The problem is that Fulfil also uses strings (according
to a specific pattern) as IDs.

This fixes that and matches that string pattern from Fulfil too.