fix: Global and local configuration overwrites
In a previous version, we changed to thread variables to store the configuration
of the FulfilApi gem. However, with this change, we introduced a problem.

When the gem was configured globally, the configuration was not shared among
all other threads. Just the one that booted the application. This meant that
processes such as Sidekiq didn't have access to the configuration and failed.

This commit fixes that problem. While keeping things thread-safe.