fix: Raise an non-429 HTTP error while doing a batch request
We automatically retry finding a batch whenever we're encountering
a 429 HTTP error (too many requests). However, if it's not a 429 HTTP
status code, we should have re-raised the original exception.

Without this re-raising in place, batch requests will hammer the Fulfil
API until the end of history. It was the root cause of the unauthorized
error not being raised for DBrand, and it was a problem while fetching
data for Abbott Lyon.

This commit fixes it by automatically re-raising an exception whenever
it's not a 429 HTTP status code.